Hosts can now kick players from the lobby

+++ src/components/GameComponent.tsx
@@ -98,6 +98,7 @@ function GameComponent(props: IGameProps) {
     let [trigger, setTrigger] = useState(false);
     let [answered, setAnswered] = useState(false);
 
+
     let gameUseRef = useRef(game);
     let currentPlayerUseRef = useRef(currentPlayer);
     let gameDocUseRef = useRef(gameDoc);
@@ -138,19 +139,27 @@ function GameComponent(props: IGameProps) {
               //@ts-ignore
               let playersDocArr = await getPlayers(props.currentGameId, playersRef);
               let playersArr : Player[] = [];
-              //@ts-ignore
+
+              let playerNotKicked = false;
               playersDocArr.forEach(player => {
                 // console.log('Player:', player);
                 playersArr.push(player);
+                
                 //@ts-ignore
                 if (player.name == props.currentUser?.username) {
-                  // console.log('Current Player:', player)
+                  console.log('Current Player:', player)
                   setCurrentPlayer(player);
                   currentPlayerUseRef.current = player;  // This is what actually works for deleting player later
                   setPlayerID(player.id);
+                  playerNotKicked = true;
                 }
                 // else console.log('ABORT: NOT THE SAME PLAYER:', player)
               })
+              // Player has been kicked if their player data does not exist in db
+              if (!playerNotKicked) {
+                console.log('Player is not in player list, must be kicked')
+                history.push('/join-game');
+              }
 
               let newGame : GameState = {
                 id: props.currentGameId as string,
@@ -188,8 +197,8 @@ function GameComponent(props: IGameProps) {
           console.log('UNMOUNTING GAME COMPONENT');
           console.log('Game in Return', gameUseRef.current);
 
-          // If player is last one in lobby, delete game          
-          if (gameUseRef.current?.players.length == 1) {
+          // If player is last one in lobby, delete game, but if player is not in list, do not delete     
+          if (gameUseRef.current?.players.length == 1 && gameUseRef.current.players.some(temp => temp.name == props.currentUser?.username)) {
             // console.log('Time to delete!');            
             firestore.deleteDoc(firestore.doc(gamesRef, `${props.currentGameId}`));
           } else {
@@ -333,11 +342,8 @@ function GameComponent(props: IGameProps) {
      */
     async function clearAnswers() {
       let playersRef = firestore.collection(gamesRef, `${props.currentGameId}/players`);
-      //@ts-ignore
       let playersDocArr = await firestore.getDocs(playersRef)
       playersDocArr.forEach(async player => {
-        //@ts-ignore
-
         let playerRef = firestore.doc(gamesRef, `${props.currentGameId}/players/${player.id}`)
         await firestore.updateDoc(playerRef, 'answered', false);
           
@@ -363,6 +369,33 @@ function GameComponent(props: IGameProps) {
     }
 
     /**
+     *  Host can kick a player from the lobby
+     *  Very unoptimized (just copied from clearAnswers)
+     */
+    async function kickPlayer(player : Player) {
+      console.log(player)
+      let playersRef = firestore.collection(gamesRef, `${props.currentGameId}/players`);      
+      let playersDocArr = await firestore.getDocs(playersRef)
+      playersDocArr.forEach(async loopPlayer => {
+        if (loopPlayer.id == player.id) {
+          console.log('Need to delete this bitch')
+          let playerRef = firestore.doc(gamesRef, `${props.currentGameId}/players/${loopPlayer.id}`)
+          // console.log(playerRef)
+          firestore.deleteDoc(playerRef);
+
+          // Send trigger update to firestore
+          let temp = await firestore.doc(gamesRef, `${props.currentGameId}`);
+          let gameDoc = await firestore.getDoc(temp);
+          //@ts-ignore
+          console.log(gameDoc['_document']['data']['value']['mapValue']['fields']['trigger'].booleanValue)
+          //@ts-ignore
+          await firestore.updateDoc(temp, 'trigger', !gameDoc['_document']['data']['value']['mapValue']['fields']['trigger'].booleanValue)
+        }
+          
+      })
+    }
+
+    /**
      *  We want to display a question field, an answer+submit field, and a list of players.
      *  Game States:
      *    - 0: Waiting to start
@@ -381,8 +414,8 @@ function GameComponent(props: IGameProps) {
                 {console.log('GAME RERENDER: ', game)}
                 {console.log('Rerendered: ', props.currentGameId, game.match_state)}              
                 {/* Player List */}
-                {/* {console.log('Players AND game in return line 187', game?.players, game)} */}
-                <PlayersComponent key={1} players={game?.players} user={props.currentUser} />
+                {console.log('Players AND game in return line 187', game?.players, game)}
+                <PlayersComponent key={1} players={game?.players} user={props.currentUser} host={game.host} kickPlayer={kickPlayer}/>
 
                 {/* If game state changes to 2, start timer, set game state to 1 when timer ends */}
                 {
@@ -555,6 +588,12 @@ function PlayersComponent(props: any) {
                                   {console.log("user" + (player.name == props.user.username), player.name, props.user.username)}
                                   {/* @ts-ignore */}
                                   {player.name} | {player.points} points
+                                  {
+                                    // Host player has Kick Player buttons attached to other players
+                                    (props.user.username == props.host && player.name != props.user.username) ?
+                                    <Button onClick={() => props.kickPlayer(player)}>Kick Player</Button>
+                                    : <></>
+                                  }
                                 </td>
                             </tr>
                           })}
